Move IceFuture.py to main directory
