--- python/python/Ice/Py3/__init__.py
@@ -1,3 +1,0 @@
-#
-# Copyright (c) ZeroC, Inc. All rights reserved.
-#
